+++ Project01 - Super Squish Sisters/BigCreate.cpp
@@ -31,7 +31,7 @@ void BigCreate::SetAttributes(string line, vector<Character*>* characters) const
 //    cout << name << " | " << std::stod(weight)<< " | " << speed<< " | "<<std::stoi(hair_length)<< " | "<<race<< " | "<<skill<<endl;
 //    cout << "**********" << endl;
     
-    Big * b = new Big(name, std::stod(weight), speed, std::stoi(hair_length), race, skill);
+    Big * b = new Big(name, std::stod(weight), speed, std::stod(hair_length), race, skill);
     characters->push_back(b);
     
 //    Big b = Big(name, std::stod(weight), speed, std::stoi(hair_length), race, skill);
