+++ Project01 - Super Squish Sisters/Medium.cpp
@@ -21,7 +21,7 @@ using std::endl;
 using std::left;
 using std::setw;
 
-Medium::Medium(string name, double weight, string speed, int hair_length, string race, string melee_weapon):Character(name,weight,speed,hair_length,race) {
+Medium::Medium(string name, double weight, string speed, double hair_length, string race, string melee_weapon):Character(name,weight,speed,hair_length,race) {
     m_meleeWeapon = melee_weapon;
 }
