+++ Project01 - Super Squish Sisters/Small.h
@@ -22,7 +22,7 @@ using std::string;
 
 class Small: public Character{
 public:
-    Small(string name, double weight, string speed, int hair_length, string race, bool gun_equipped);
+    Small(string name, double weight, string speed, double hair_length, string race, bool gun_equipped);
 
     virtual ~Small();
